+++ griptape/tasks/toolkit_task.py
@@ -68,7 +68,7 @@ class ToolkitTask(PromptTask, ActionsSubtaskOriginMixin):
     @property
     def prompt_stack(self) -> PromptStack:
         stack = PromptStack(tools=self.tools)
-        memory = self.structure.conversation_memory if self.structure is not None else None
+        memory = self.conversation_memory
 
         stack.add_system_message(self.generate_system_template(self))